Fix conversion warning

+++ Python/pystate.c
@@ -2539,7 +2539,7 @@ tstate_mimalloc_bind(PyThreadState *tstate)
     tld->segments.abandoned = &tstate->interp->mimalloc.abandoned_pool;
 
     // Initialize each heap
-    for (size_t i = 0; i < _Py_MIMALLOC_HEAP_COUNT; i++) {
+    for (uint8_t i = 0; i < _Py_MIMALLOC_HEAP_COUNT; i++) {
         _mi_heap_init_ex(&mts->heaps[i], tld, _mi_arena_id_none(), false, i);
     }
